src/test/java/com/dominity/ing/onlinegame/service/OnlineGameServiceTest.java
package com.dominity.ing.onlinegame.service;

import org.junit.jupiter.api.Test;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.nio.charset.StandardCharsets;

import static org.assertj.core.api.Assertions.assertThat;

class OnlineGameServiceTest {
    @Test
    void testExample1() throws IOException {
        // given
        var requestJson = "{\"groupCount\":6,\"clans\":[{\"numberOfPlayers\":4,\"points\":50},{\"numberOfPlayers\":2,\"points\":70},{\"numberOfPlayers\":6,\"points\":60},{\"numberOfPlayers\":1,\"points\":15},{\"numberOfPlayers\":5,\"points\":40},{\"numberOfPlayers\":3,\"points\":45},{\"numberOfPlayers\":1,\"points\":12},{\"numberOfPlayers\":4,\"points\":40}]}";
        var responseJson = "[[{\"numberOfPlayers\":2,\"points\":70},{\"numberOfPlayers\":4,\"points\":50}],[{\"numberOfPlayers\":6,\"points\":60}],[{\"numberOfPlayers\":3,\"points\":45},{\"numberOfPlayers\":1,\"points\":15},{\"numberOfPlayers\":1,\"points\":12}],[{\"numberOfPlayers\":4,\"points\":40}],[{\"numberOfPlayers\":5,\"points\":40}]]";
        var outputStream = new ByteArrayOutputStream();
        // when
        new OnlineGameService().groupClans(new ByteArrayInputStream(requestJson.getBytes(StandardCharsets.UTF_8)), outputStream);
        // then
        assertThat(outputStream.toByteArray()).isEqualTo(responseJson.getBytes(StandardCharsets.UTF_8));
    }
}
